# ArchitectureRefactoringHelper/Repository/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Repository, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/outputs/{id}": {
      "get": {
        "tags": [
          "ApproachOutput"
        ],
        "operationId": "GetApproachOutput",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApproachOutput"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outputs/architectures": {
      "get": {
        "tags": [
          "ApproachOutput"
        ],
        "operationId": "ListArchitectures",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Architecture"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachOutput"
        ],
        "operationId": "AddArchitecture",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Architecture"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Architecture"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Architecture"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Architecture"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outputs/architectures/{name}": {
      "delete": {
        "tags": [
          "ApproachOutput"
        ],
        "operationId": "DeleteArchitecture",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/outputs/service-types": {
      "get": {
        "tags": [
          "ApproachOutput"
        ],
        "operationId": "ListServiceTypes",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceType"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachOutput"
        ],
        "operationId": "AddServiceType",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceType"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceType"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ServiceType"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceType"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outputs/service-types/{name}": {
      "delete": {
        "tags": [
          "ApproachOutput"
        ],
        "operationId": "DeleteServiceType",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/v1/outputs/representations": {
      "get": {
        "tags": [
          "ApproachOutput"
        ],
        "operationId": "ListRepresentations",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Representation"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ApproachOutput"
        ],
        "operationId": "AddRepresentation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Representation"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Representation"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Representation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Representation"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/outputs/representations/{name}": {
      "delete": {
        "tags": [
          "ApproachOutput"
        ],
        "operationId": "DeleteRepresentation",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/v1/approaches/{id}/outputs/{outputId}": {
      "delete": {
        "tags": [
          "RefactoringApproach"
        ],
        "operationId": "RemoveOutput",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "outputId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproachOutput": {
        "required": [
          "architecture",
          "serviceType"
        ],
        "type": "object",
        "properties": {
          "approachOutputId": {
            "type": "integer",
            "format": "int32"
          },
          "architecture": {
            "$ref": "#/components/schemas/Architecture"
          },
          "serviceType": {
            "$ref": "#/components/schemas/ServiceType"
          }
        },
        "additionalProperties": false
      },
...
      "Architecture": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Representation": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ServiceType": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}